requirements.txt
alembic==1.14.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.5.2
attrs==25.3.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.8.2
cachetools==5.5.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
deprecation==2.1.0
distlib==0.4.0
distro==1.9.0
exceptiongroup==1.3.1
filelock==3.16.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fpdf==1.7.2
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.1.0
google-api-core==2.28.1
google-auth==2.43.0
google-generativeai==0.1.0rc1
googleapis-common-protos==1.72.0
gotrue==2.9.2
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.62.3
h11==0.16.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.9
httpx==0.27.2
hyperframe==6.0.1
idna==3.11
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.42.1
numpy==1.24.4
openai==2.2.0
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pipenv==2024.4.1
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
playwright==1.48.0
pluggy==1.5.0
postgrest==0.16.11
proto-plus==1.26.1
protobuf==4.25.8
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pyee==12.0.0
Pygments==2.19.2
PyJWT==2.9.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
realtime==1.0.6
referencing==0.35.1
requests==2.32.4
rich==13.9.4
rpds-py==0.20.1
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.49
storage3==0.7.7
streamlit==1.40.1
StrEnum==0.4.15
supabase==2.6.0
supafunc==0.5.1
tenacity==9.0.0
toml==0.10.2
tomli==2.4.0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
virtualenv==20.35.4
watchdog==4.0.2
websockets==12.0
Werkzeug==3.0.6
zipp==3.20.2
rich==14.1.*